Fix bug about Float Not work in BuyIn Form

+++ src/Entity/BuyIn.php
@@ -18,7 +18,7 @@ class BuyIn
     private $id;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $value;
 
@@ -33,12 +33,12 @@ class BuyIn
         return $this->id;
     }
 
-    public function getValue(): ?int
+    public function getValue(): ?float
     {
         return $this->value;
     }
 
-    public function setValue(int $value): self
+    public function setValue(float $value): self
     {
         $this->value = $value;
